fix: corrige salvar matricula

+++ app/src/main/java/br/edu/fateczl/presencaesportiva/DAO/turmaDAO.java
@@ -8,7 +8,7 @@ import br.edu.fateczl.presencaesportiva.model.Turma;
 public interface turmaDAO {
     void cadastrar(Turma a);
     void apagar(Turma f);
-    void atualizar(int id, Turma f);
+    void atualizar(long id, Turma f);
     List<Turma> pesquisarPorNome(String nome);
 }
